requirements.txt
discord.py>=2.3.2
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.9.1
newsapi-python>=0.2.7
deep-translator>=1.11.4
pytz>=2023.3
beautifulsoup4>=4.12.2
feedparser>=6.0.10
Pillow>=10.1.0
instaloader>=4.10.3
